job-market-pipeline/pyproject.toml
[project]
name = "job-market-pipeline"
version = "0.1.0"
description = "Production-ready pipeline for LinkedIn job market data"
dependencies = [
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-bigquery>=3.17.0",
    "dbt-bigquery>=1.7.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
managed = true

[tool.dbt]
project-dir = "dbt"
profiles-dir = "dbt"
